requirements.txt
python-socketio
eventlet==0.30.2
gunicorn==20.1.0
SQLAlchemy==1.4.45